use get_pagerduty_service_id for measure operations
