+++ tests/manage/monitoring/pagerduty/test_deployment_status.py
@@ -19,20 +19,6 @@ from ocs_ci.utility import pagerduty
 log = logging.getLogger(__name__)
 
 
-def get_pagerduty_service_id():
-    """
-    Get the pagerduty service id in the current run
-
-    Returns:
-        str: The pagerduty service id in the current run
-
-    """
-    if config.multicluster:
-        return config.default_cluster_ctx.RUN["pagerduty_service_id"]
-    else:
-        return config.RUN["pagerduty_service_id"]
-
-
 @blue_squad
 @tier4
 @tier4c
@@ -64,7 +50,7 @@ def test_ceph_manager_stopped_pd(measure_stop_ceph_mgr):
         api.check_incident_cleared(
             summary=target_label,
             measure_end_time=measure_stop_ceph_mgr.get("stop"),
-            pagerduty_service_ids=[get_pagerduty_service_id()],
+            pagerduty_service_ids=[pagerduty.get_pagerduty_service_id()],
         )
 
 
@@ -99,12 +85,12 @@ def test_ceph_osd_stopped_pd(measure_stop_ceph_osd):
     api.check_incident_cleared(
         summary=constants.ALERT_OSDDISKUNAVAILABLE,
         measure_end_time=measure_stop_ceph_osd.get("stop"),
-        pagerduty_service_ids=[get_pagerduty_service_id()],
+        pagerduty_service_ids=[pagerduty.get_pagerduty_service_id()],
     )
     api.check_incident_cleared(
         summary=constants.ALERT_OSDDISKNOTRESPONDING,
         measure_end_time=measure_stop_ceph_osd.get("stop"),
-        pagerduty_service_ids=[get_pagerduty_service_id()],
+        pagerduty_service_ids=[pagerduty.get_pagerduty_service_id()],
     )
 
 
@@ -139,7 +125,7 @@ def test_stop_worker_nodes_pd(measure_stop_worker_nodes):
         api.check_incident_cleared(
             summary=target_label,
             measure_end_time=measure_stop_worker_nodes.get("stop"),
-            pagerduty_service_ids=[get_pagerduty_service_id()],
+            pagerduty_service_ids=[pagerduty.get_pagerduty_service_id()],
         )
 
 
@@ -178,7 +164,7 @@ def test_ceph_monitor_stopped_pd(measure_stop_ceph_mon):
             summary=target_label,
             measure_end_time=measure_stop_ceph_mon.get("stop"),
             time_min=time_min,
-            pagerduty_service_ids=[get_pagerduty_service_id()],
+            pagerduty_service_ids=[pagerduty.get_pagerduty_service_id()],
         )
 
 
@@ -211,5 +197,5 @@ def test_ceph_mons_quorum_lost_pd(measure_stop_ceph_mon):
     api.check_incident_cleared(
         summary=target_label,
         measure_end_time=measure_stop_ceph_mon.get("stop"),
-        pagerduty_service_ids=[get_pagerduty_service_id()],
+        pagerduty_service_ids=[pagerduty.get_pagerduty_service_id()],
     )
